# backend/docs/openapi.json
{
    "components": {
        "schemas": {
...
            "badges.Badge": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "course_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "issued_at": {
                        "type": "string"
                    },
                    "revoked_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "badges.BadgeResponse": {
                "properties": {
                    "code": {
                        "example": "a1b2c3d4e5f6",
                        "type": "string"
                    },
                    "course_id": {
                        "example": 1,
                        "type": "integer"
                    },
                    "issued_at": {
                        "example": "2026-09-10T02:00:00Z",
                        "type": "string"
                    },
                    "valid": {
                        "example": true,
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "badges.MyBadgesResponse": {
                "properties": {
                    "badges": {
                        "items": {
                            "$ref": "#/components/schemas/badges.Badge"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "example": 1,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "courses.Matricula": {
                "properties": {
                    "course_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "fecha_inscripcion": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "inscrito": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "string"
                    },
                    "student_id": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "enroll.EnrollRequest": {
                "properties": {
                    "course_id": {
                        "example": 1,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "enroll.EnrollmentsResponse": {
                "properties": {
                    "enrollments": {
                        "items": {
                            "$ref": "#/components/schemas/courses.Matricula"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total": {
                        "example": 3,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "enroll.MatriculaEnvelope": {
                "properties": {
                    "enrollment": {
                        "$ref": "#/components/schemas/courses.Matricula"
                    }
                },
                "type": "object"
            },
...
            "utils.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Se requieren username, email y password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "utils.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "Operación exitosa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Token JWT de sesión. Formato: `Bearer \u003ctoken\u003e`.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "API REST del monolito modular: identidad y sesiones, administración y auditoría, autoría de cursos versionados y carga multimedia con procesamiento asíncrono a HLS.\nAutenticación con JWT revocable: iniciar sesión en /api/v1/auth/login y enviar `Authorization: Bearer \u003ctoken\u003e`.",
        "title": "Plataforma MOOC - API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1/badges": {
            "get": {
                "description": "Retorna las insignias emitidas al estudiante (incluye IDs para gestión).",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/badges.MyBadgesResponse"
                                }
                            }
                        },
                        "description": "Insignias"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mis insignias",
                "tags": [
                    "Insignias"
                ]
            }
        },
        "/api/v1/badges/{code}": {
            "get": {
                "description": "URL pública de verificación; no expone datos del estudiante.",
                "parameters": [
                    {
                        "description": "Código de la insignia",
                        "in": "path",
                        "name": "code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/badges.BadgeResponse"
                                }
                            }
                        },
                        "description": "Insignia válida"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrada o revocada"
                    }
                },
                "summary": "Verificar insignia",
                "tags": [
                    "Insignias"
                ]
            }
        },
        "/api/v1/badges/{id}": {
            "delete": {
                "description": "Marca RevokedAt; la URL pública deja de validar.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID de la insignia",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Insignia revocada"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Revocar insignia",
                "tags": [
                    "Insignias"
                ]
            }
        },
...
        "/api/v1/enrollments": {
            "get": {
                "description": "Lista cursos con inscripción del estudiante, incluyendo retirados (progreso conservado).",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/enroll.EnrollmentsResponse"
                                }
                            }
                        },
                        "description": "Inscripciones"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mis inscripciones",
                "tags": [
                    "Inscripción"
                ]
            },
            "post": {
                "description": "Crea o reactiva la inscripción; re-POST no duplica ni borra progreso.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/enroll.EnrollRequest",
                                "summary": "request",
                                "description": "Curso"
                            }
                        }
                    },
                    "description": "Curso",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/enroll.MatriculaEnvelope"
                                }
                            }
                        },
                        "description": "Ya inscrito"
                    },
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/enroll.MatriculaEnvelope"
                                }
                            }
                        },
                        "description": "Inscrito"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Curso no publicado"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Inscribirse",
                "tags": [
                    "Inscripción"
                ]
            }
        },
        "/api/v1/enrollments/{course_id}": {
            "delete": {
                "description": "Marca Inscrito=false; el progreso y las notas se conservan para reinscripción.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del curso",
                        "in": "path",
                        "name": "course_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Retiro registrado"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Retirarse",
                "tags": [
                    "Inscripción"
                ]
            }
        },
        "/api/v1/enrollments/{course_id}/reenroll": {
            "post": {
                "description": "Reactiva la inscripción sin tocar progreso ni notas.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del curso",
                        "in": "path",
                        "name": "course_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/enroll.MatriculaEnvelope"
                                }
                            }
                        },
                        "description": "Reinscrito"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Curso no publicado"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Reinscribirse",
                "tags": [
                    "Inscripción"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080"
        }
    ]
}